api_traducteur/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.5
google-pasta==0.2.0
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
keras==3.4.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mysql-connector-python==8.4.0
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.5
packaging==24.0
protobuf==4.25.3
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
sacremoses==0.1.1
safetensors==0.4.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
